Add ability to define timelines in source code.
This also makes all effect parameters public in order to be able to change them after emplacing an effect to the timeline. While this is problematic because changing effect parameters after they have been added to a ParticleRenderer is not supported, I think this is ok because the user really shouldn't hold the raw pointers for very long and can't access the timeline again after it has been added to a ParticleRenderer (assuming he didn't do any raw pointer hacks).