Calculate ParticleData on the GPU, OpenGL Debug stuff
This implements calculating the ParticleData (rgc->hsv, foregroundMask, ...) on the GPU using transform feedback. This serves as a preparation for mirroring the viewer multiple times on the screen and using this as input for the particle effects.

OpenGL debug stuff includes: DebugMessageCallback from OpenGL implementation (requires OpenGL 4.2+) for builds with debug info, saving textures to file, code to read ParticleData